+++ pyproject.toml
@@ -10,6 +10,7 @@ readme = "README.md"
 python = ">=3.11"
 requests = "^2.31.0"
 pandas = "^2.2.1"
+aiohttp = "^3.9.4"
 
 [tool.poetry.group.dev.dependencies]
 pre-commit = "^3.5.0"
