--- src/aiod_sdk/endpoints/case_studies.py
@@ -1,3 +1,0 @@
-from . import calls
-
-get_list, counts, get_asset = calls.wrap_common_calls(asset_type="case_studies")
