--- src/aiod_sdk/endpoints/educational_resources.py
@@ -1,5 +1,0 @@
-from . import calls
-
-get_list, counts, get_asset = calls.wrap_common_calls(
-    asset_type="educational_resources"
-)
