+++ src/aiod_sdk/resources/educational_resources.py
@@ -1,0 +1,8 @@
+from aiod_sdk.calls import calls
+
+get_list, counts, get_asset = calls.wrap_common_calls(
+    asset_type="educational_resources"
+)
+get_asset_async, get_list_async = calls.wrap_common_async_calls(
+    asset_type="educational_resources"
+)
