+++ src/aiod_sdk/resources/ml_models.py
@@ -1,0 +1,6 @@
+from aiod_sdk.calls import calls
+
+get_list, counts, get_asset = calls.wrap_common_calls(asset_type="ml_models")
+get_asset_async, get_list_async = calls.wrap_common_async_calls(
+    asset_type="experiments"
+)
