+++ src/aiod_sdk/resources/persons.py
@@ -1,0 +1,4 @@
+from aiod_sdk.calls import calls
+
+get_list, counts, get_asset = calls.wrap_common_calls(asset_type="persons")
+get_asset_async, get_list_async = calls.wrap_common_async_calls(asset_type="persons")
